+++ doc/misc/changelog.md
@@ -41,6 +41,10 @@ Behavior changes
 * Calling {func}`fwdpy11.infinite_sites` during a simulation now raises `RuntimeError`.
   {pr}`820`
   {issue}`769`
+* Models imported from `demes` now start the forward-time portion of the model 1 (one) generation before the most ancient end time of an ancestral deme.
+  {pr}`818`
+  {user}`apragsdale`
+  {user}`molpopgen`
 
 New features